+++ CHANGES.md
@@ -4,6 +4,9 @@
 
 ## Changes
 
+* Remove dependency on libabac
+([#467](https://github.com/GENI-NSF/geni-ch/issues/467))
+
 ## Installation Notes
 
 # [Release 2.11.1](https://github.com/GENI-NSF/geni-ch/milestones/2.11.1)