+++ CHANGES.md
@@ -6,6 +6,8 @@
 
 * Provide project credentials via SA.get_credentials
 ([#466](https://github.com/GENI-NSF/geni-ch/issues/466))
+* Remove dependency on libabac
+([#467](https://github.com/GENI-NSF/geni-ch/issues/467))
 
 ## Installation Notes