+++ CHANGES.md
@@ -4,6 +4,8 @@
 
 ## Changes
 
+* Provide project credentials via SA.get_credentials
+([#466](https://github.com/GENI-NSF/geni-ch/issues/466))
 * Remove dependency on libabac
 ([#467](https://github.com/GENI-NSF/geni-ch/issues/467))